Update Admin UI for multiple reimbursements per customer return
Tweaks the admin ui a bit so that multiple reimbursements per customer return are displayed
correctly.

We originally only expected a single reimbursement per customer return. But with expedited exchanges
there may be a reimbursement that happens immediately (the exchange shipment) for any exchange items
in the RMA and another reimbursement that happens later when the return items arrive -- when the
other non-exchange items are reimbursed for money.  This new display logic is more robust in any
case.

cherry-picked from 0a2f382885c7766996ece206567da10587c35e26 in spree master